fix pixel events on Single Product page
