+++ includes/Tracking/PixelTrackingService.php
@@ -107,6 +107,11 @@ final class PixelTrackingService implements ServiceStatusInterface {
 			array( $this->tracker, 'track_purchase_event' ),
 			11
 		);
+
+		add_action(
+			'woocommerce_after_add_to_cart_quantity',
+			array( $this, 'render_event_id_field' )
+		);
 	}
 
 	/**
@@ -259,5 +264,34 @@ final class PixelTrackingService implements ServiceStatusInterface {
 	 */
 	public function filter_page_view_event_data( &$tracking_data ): void {
 		$tracking_data['PAGE_VIEW'] = true;
+	}
+
+	/**
+	 * Outputs a hidden input field for the Event ID on single product pages.
+	 *
+	 * This is used to inject a unique UUID per Add to Cart action, enabling
+	 * deduplication between Pixel and CAPI events.
+	 *
+	 * Also injects inline JavaScript that generates a `window.crypto.randomUUID()` and
+	 * assigns it to the hidden field.
+	 *
+	 * @since 0.1.0
+	 *
+	 * @return void
+	 */
+	public function render_event_id_field(): void {
+		$attr = Helper::with_prefix( 'event_id' );
+
+		printf(
+			'<input type="hidden" name="%1$s" value="" />',
+			esc_attr( $attr ),
+		);
+
+		wp_print_inline_script_tag(
+			sprintf(
+				'document.querySelector("[name=%s]").value = window.crypto.randomUUID()',
+				esc_attr( $attr )
+			)
+		);
 	}
 }
